:memo: add docstrings (part 1) #sdk/k8s-operator
