+++ sources/sdk/k8s-operator/src/kube-interface.js
@@ -1,7 +1,15 @@
+/**
+ * Wrapper for Kubernetes API Server client.
+ * @module kube-interface
+ */
+
 const Request = require('kubernetes-client/backends/request')
 const kubernetes = require('kubernetes-client')
 
-
+/**
+ * Domain specific error.
+ * @class KubeError
+ */
 class KubeError extends Error {
   constructor(msg, details = null) {
     super(msg)
@@ -80,7 +88,22 @@ const response = {
   }
 }
 
+/**
+ * Wrapper object for Kubernetes API Server client
+ * @class KubeInterface
+ */
 class KubeInterface {
+  /**
+   * @typedef {Object} KubeInterfaceConfiguration
+   * @property {Array} crds List of Custom Resource Definitions to add
+   * @property {Boolean} createCRDS Create CRDs unknown by the server if enabled
+   * @property {Object} [config] KubeConfig to use (or `null`)
+   */
+
+  /**
+   * Create a new Kubernetes client
+   * @param {KubeInterfaceConfiguration} params
+   */
   constructor({ crds = [], createCRDs = true, config = null }) {
     if (config !== null) {
       this.config = config
@@ -96,6 +119,16 @@ class KubeInterface {
     this.createCRDs = createCRDs
   }
 
+  /**
+   * Load the API definition from the server.
+   *
+   * @returns {Promise<void>}
+   *
+   * This is located either at /openapi/v2 or /swagger.json
+   *
+   * NB: if the `createCRDs` flag is set, every CRDs from the `crds` array
+   *     unknown to the server will be created.
+   */
   async load() {
     await this.client.loadSpec()
 
@@ -123,6 +156,14 @@ class KubeInterface {
     this.crds.map(this.client.addCustomResourceDefinition.bind(this.client))
   }
 
+  /**
+   * Create resources.
+   *
+   * NB: Resources are created in parallel.
+   *
+   * @param  {...Object} resources Resource to create
+   * @returns {Promise<Array<Object>>} Resolve to the list of created ressources or reject with the first thrown error
+   */
   async create(...resources) {
     return await Promise.all(resources
       .map(resource => ({
@@ -147,6 +188,20 @@ class KubeInterface {
     )
   }
 
+  /**
+   * @typedef {Object} ListRequestInfo
+   * @property {String} apiVersion Kubernetes resource API version
+   * @property {String} kind Kubernetes resource kind
+   * @property {String} [namespace] Kubernetes resource namespace
+   * @property {String} [labels] Kubernetes resource label filter
+   */
+
+  /**
+   * List resources matching the supplied parameters.
+   *
+   * @param {ListRequestInfo} params
+   * @returns {Promise<Array<Object>>} Matching Kubernetes resources
+   */
   async list({ apiVersion, kind, namespace, labels }) {
     const { apiGroup, resourceVersion } = parseApiVersion(apiVersion)
     const endpoint = getEndpoint(this.client, {
@@ -162,6 +217,20 @@ class KubeInterface {
     )
   }
 
+  /**
+   * @typedef {Object} GetRequestInfo
+   * @property {String} apiVersion Kubernetes resource API version
+   * @property {String} kind Kubernetes resource kind
+   * @property {String} [namespace] Kubernetes resource namespace
+   * @property {String} name Kubernetes resource name
+   */
+
+  /**
+   * Get a single resource by name.
+   *
+   * @param {GetRequestInfo} params
+   * @returns {Promise<Object>} Matching Kubernetes resource
+   */
   async get({ apiVersion, kind, namespace, name }) {
     const { apiGroup, resourceVersion } = parseApiVersion(apiVersion)
     const endpoint = getEndpoint(this.client, {
@@ -175,6 +244,21 @@ class KubeInterface {
     return response.unwrap(await endpoint.get())
   }
 
+  /**
+   * @typedef {Object} WatchRequestInfo
+   * @property {String} apiVersion Kubernetes resource API version
+   * @property {String} kind Kubernetes resource kind
+   * @property {String} [namespace] Kubernetes resource namespace
+   * @property {String} [name] Kubernetes resource name
+   */
+
+  /**
+   * Watch Kubernetes resources.
+   * A stream object will be returned, and Kubernetes events will be received on the 'data' stream event
+   *
+   * @param {WatchRequestInfo} params
+   * @returns {Promise<stream.Readable>}
+   */
   async watch({ apiVersion, kind, namespace, name }) {
     const { apiGroup, resourceVersion } = parseApiVersion(apiVersion)
     const endpoint = getEndpoint(this.client, {
@@ -188,6 +272,29 @@ class KubeInterface {
     return await endpoint.getObjectStream()
   }
 
+  /**
+   * @typedef {Object} WaitConditionRequestInfo
+   * @property {String} apiVersion Kubernetes resource API version
+   * @property {String} kind Kubernetes resource kind
+   * @property {String} [namespace] Kubernetes resource namespace
+   * @property {String} [name] Kubernetes resource name
+   * @property {Function} [callback] Condition check callback
+   */
+
+  /**
+   * @typedef {Object} WaitConditionCallbackResult
+   * @property {Boolean} condition Wether the condition was met or not
+   * @property {any} res Result to return
+   */
+
+  /**
+   * Watch resources until a condition is true.
+   * The callback must return a Promise resolving to an object of
+   * type {@link WaitConditionCallbackResult}
+   *
+   * @param {WaitConditionRequestInfo} params
+   * @returns {Promise<any>} Result returned by the first callback that met its condition
+   */
   async waitCondition({
     apiVersion,
     kind,
@@ -220,6 +327,21 @@ class KubeInterface {
     return result
   }
 
+  /**
+   * @typedef PatchRequestInfo
+   * @property {String} apiVersion Kubernetes resource API version
+   * @property {String} kind Kubernetes resource kind
+   * @property {String} [namespace] Kubernetes resource namespace
+   * @property {String} name Kubernetes resource name
+   * @property {Object} patch Kubernetes Resource Merge Patch
+   */
+
+  /**
+   * Patch a Kubernetes resource with a {@link https://tools.ietf.org/html/rfc7386|JSON Merge Patch}.
+   *
+   * @param {PatchRequestInfo} params
+   * @returns {Promise<Object>} Patched Kubernetes resource
+   */
   async patch({ apiVersion, kind, namespace, name, patch }) {
     const { apiGroup, resourceVersion } = parseApiVersion(apiVersion)
     const endpoint = getEndpoint(this.client, {
@@ -238,6 +360,20 @@ class KubeInterface {
     }))
   }
 
+  /**
+   * @typedef {Object} DeleteRequestInfo
+   * @property {String} apiVersion Kubernetes resource API version
+   * @property {String} kind Kubernetes resource kind
+   * @property {String} [namespace] Kubernetes resource namespace
+   * @property {String} name Kubernetes resource name
+   */
+
+  /**
+   * Delete a single Kubernetes resources
+   *
+   * @param {DeleteRequestInfo} params
+   * @returns {Promise<Object>} Deleted Kubernetes resource
+   */
   async delete({ apiVersion, kind, namespace, name }) {
     const { apiGroup, resourceVersion } = parseApiVersion(apiVersion)
     const endpoint = getEndpoint(this.client, {
@@ -251,6 +387,20 @@ class KubeInterface {
     return response.unwrap(await endpoint.delete())
   }
 
+  /**
+   * @typedef ExecRequestInfo
+   * @property {String} namespace Kubernetes Pod namespace
+   * @property {String} name Kubernetes Pod name
+   * @property {String} command Command to execute inside the Pod's container
+   * @property {String} [container] Pod container name (mandatory if more than one container)
+   */
+
+  /**
+   * Execute a command in a Pod's container.
+   *
+   * @param {ExecRequestInfo} params
+   * @returns {Promise<String>} Command output
+   */
   async exec({ namespace, name, command, container }) {
     const { apiGroup, resourceVersion } = parseApiVersion('v1')
     const endpoint = getEndpoint(this.client, {
@@ -270,6 +420,19 @@ class KubeInterface {
     }))
   }
 
+  /**
+   * @typedef LogRequestInfo
+   * @property {String} namespace Kubernetes Pod namespace
+   * @property {String} name Kubernetes Pod name
+   * @property {String} [container] Pod container name (mandatory if more than one container)
+   */
+
+  /**
+   * Get Pod's container logs
+   *
+   * @param {LogRequestInfo} params
+   * @returns {Promise<String>} Pod's container logs
+   */
   async logs({ namespace, name, container }) {
     const { apiGroup, resourceVersion } = parseApiVersion('v1')
     const endpoint = getEndpoint(this.client, {
