+++ sources/sdk/k8s-operator/src/web-service.js
@@ -1,6 +1,20 @@
+/**
+ * Handle Kubernetes Pod's lifecycle.
+ * @module web-service
+ */
+
 const { createTerminus } = require('@godaddy/terminus')
 
+/**
+ * Operator Web Service LifeCycle manager
+ * @class WebService
+ */
 class WebService {
+  /**
+   * Create a new WebService
+   * @param {Operator} operator Operator owning the web service
+   * @param {ServerFactory} serverFactory HTTP(S) server factory
+   */
   constructor(operator, serverFactory) {
     this.operator = operator
     this.servers = serverFactory.make(this.operator.webapp)
@@ -20,6 +34,10 @@ class WebService {
     }))
   }
 
+  /**
+   * Callback called before openning the HTTP(S) port(s).
+   * @returns {Promise<void>}
+   */
   async beforeListen() {
     await this.operator.kubectl.load()
     await this.operator.initialize()
@@ -29,26 +47,56 @@ class WebService {
     )
   }
 
+  /**
+   * Callback called before shutdown triggered by Kubernetes Pod's restart
+   * @returns {Promise<void>}
+   */
   async beforeShutdown() {
     this.cancelScopes.map(cancelScope => cancelScope.cancel())
   }
+
+  /**
+   * Callback called when a shutdown was triggered by the Kubernetes Pod
+   * @returns {Promise<void>}
+   */
 
   async shutdownRequested() {
     console.log('Shutdown requested')
   }
 
+  /**
+   * Callback called after HTTP(S) server(s) shutdown
+   * @returns {Promise<void>}
+   */
   async shutdownDone() {
     await this.operator.terminate()
   }
 
+  /**
+   * Callback called when the HTTP(S) server(s) shutdown failed
+   *
+   * @param {Error} err Error thrown during shutdown
+   * @returns {Promise<void>}
+   */
   async shutdownFailed(err) {
     console.error(err)
   }
 
+  /**
+   * Callback called when the terminus library wants to log information.
+   *
+   * @param {any} msg
+   * @param {any} payload
+   * @returns {Promise<void>}
+   */
   async logger(msg, payload) {
     console.log(msg, payload)
   }
 
+  /**
+   * Open HTTP(S) port(s).
+   * @returns {Promise<void>} Resolves once all ports are open
+   */
   async listen() {
     await this.beforeListen()
